ci: cast uint64->double narrowings to clear MSVC C4244 under /W4 /WX
MSVC /W4 flags every implicit conversion of a 64-bit integer (uint64_t,
size_t, Size, Cycle, ...) or lossy 32-bit int to double/float as C4244
"possible loss of data". The GNU builds use -Wall -Wextra (no -Wconversion)
so they don't see these; the MSVC build stops at the first one, surfacing
them a few files at a time on CI.

Enumerated the full set locally with a clang scan adding
-Wimplicit-int-float-conversion / -Wfloat-conversion (which fire only on
genuinely lossy conversions — a faithful match for MSVC C4244, verified the
actual MSVC failure sites appear in it), then applied a minimal explicit
static_cast<double>/<float> at each of ~626 sites across 152 files. These are
stats/timing/throughput computations (cycles, bytes, op counts converted to
double for ratios, bandwidth, utilization, percentages). For ratio lines both
operands of the division are now cast where both were integers. No signatures,
member types, or aliases changed; casts are behavior-preserving (they make the
conversion the compiler already performed explicit).

Verified clean locally under release-werror with BOTH gcc and clang, and a
clang -Wimplicit-int-float-conversion/-Wfloat-conversion scan now reports zero
double/float narrowings in our sources.

+++ examples/basic/concurrent_execution_debug.cpp
@@ -295,9 +295,9 @@ void debug_timing_calculation() {
     // bytes / (GB/s) = bytes / (10^9 bytes/s) = bytes * 10^-9 s = bytes ns
     // At 1 GHz: cycles = bytes / bandwidth_gb_s
 
-    Cycle dma_cycles = static_cast<Cycle>(tile_size_bytes / dma_bw);
-    Cycle bm_cycles = static_cast<Cycle>(tile_size_bytes / bm_bw);
-    Cycle str_cycles = static_cast<Cycle>(tile_size_bytes / str_bw);
+    Cycle dma_cycles = static_cast<Cycle>(static_cast<double>(tile_size_bytes) / dma_bw);
+    Cycle bm_cycles = static_cast<Cycle>(static_cast<double>(tile_size_bytes) / bm_bw);
+    Cycle str_cycles = static_cast<Cycle>(static_cast<double>(tile_size_bytes) / str_bw);
 
     std::cout << "Tile size: " << tile_size_bytes << " bytes (16x16 float32)\n\n";
